move ProductData type to ~/types/shopify.ts
