+++ store/products/mutations.ts
@@ -1,6 +1,6 @@
 import { MutationTree } from 'vuex';
 import { ProductsState } from '~/types/products.store';
-import { ProductData } from '~/types/pdp.store';
+import type { ProductData } from '~/types/shopify';
 
 const mutations: MutationTree<ProductsState> = {
   setAllProducts: (state, products: ProductData[]) =>
