+++ store/products/state.ts
@@ -1,5 +1,5 @@
 import { ProductsState } from '~/types/products.store';
-import type { ProductData } from '~/types/pdp.store';
+import type { ProductData } from '~/types/shopify';
 
 const state = (): ProductsState => ({
   allProducts: [] as ProductData[],
