+++ types/pdp.store.ts
@@ -1,19 +1,4 @@
-import type {
-  ProductVariants,
-  ProductImages,
-  PriceRange,
-} from '~/types/shopify';
-
-export type ProductData = {
-  __typename: string;
-  id: string;
-  title: string;
-  description: string;
-  handle: string;
-  images: ProductImages;
-  priceRange: PriceRange;
-  variants: ProductVariants;
-};
+import type { ProductData } from '~/types/shopify';
 
 export interface PDPState {
   productData: ProductData;
