+++ types/products.store.ts
@@ -1,4 +1,4 @@
-import type { ProductData } from './pdp.store';
+import type { ProductData } from '~/types/shopify';
 
 export interface ProductsState {
   allProducts: ProductData[];
